This was worse than expected #54
There was something keeping me up with the other commit. Row counts didn't match.

I did my best but the current data is too different to the original and I don't have access to the original source files. In the new version, there are four tail numbers that no longer exist in `planes`.  However, there are now 38 more tail numbers in `planes` and 1213 more `flights` that have a tail number with an associated plane.